enforce cert chain, bind hashmap
certificate chain is verified to link to root (no self-signing);
hashmap array is constrained to blob limits;
some 64-bit compilation warnings are fixed;
rebuilt tools with MSVC 2015;
only "small" libraries were rebuilt

+++ TaggantLibrary/TaggantLibrary/src/http.c
@@ -235,7 +235,7 @@ int http_read(const char *hostName, const void *requestString, int requestLength
                     r_end = strstr(r_start, "\r\n");
                     if (r_end && r_end > r_start)
                     {
-                        *resultBufferLength = r_end - r_start;
+                        *resultBufferLength = (int) (r_end - r_start);
                         *resultBuffer = memory_alloc(*resultBufferLength + 1);
                         if (!*resultBuffer)
                         {
